requirements.txt
astroid==2.5.6
certifi==2020.12.5
chardet==4.0.0
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pylint==2.8.2
pyst2==0.5.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
